refactor: add scss feature
