+++ src/components/common.module.scss
@@ -2,17 +2,16 @@
   padding: 5px 15px;
   color: #ffffff;
   font-size: 1rem;
-  font-weight: bold;
   background: transparent;
   border: 1px solid #333333;
   cursor: pointer;
   border-radius: 5px;
   transition: 0.2s ease-in-out;
-}
 
-.button:hover {
-  border-color: #606060;
-  -webkit-box-shadow: inset 0 0 15px #ffffff50;
-  -moz-box-shadow: inset 0 0 15px #ffffff50;
-  box-shadow: inset 0 0 15px #ffffff50;
+  &:hover {
+    border-color: #606060;
+    -webkit-box-shadow: inset 0 0 15px #ffffff50;
+    -moz-box-shadow: inset 0 0 15px #ffffff50;
+    box-shadow: inset 0 0 15px #ffffff50;
+  }
 }
