+++ src/components/layout/error-trigger/ErrorTrigger.module.scss
@@ -3,10 +3,8 @@
   z-index: 1;
   top: 10px;
   left: 10px;
-}
 
-@media screen and (max-width: 900px) {
-  .errorTrigger {
+  @media screen and (max-width: 900px) {
     left: 50%;
     top: 50px;
     transform: translateX(-50%);
