+++ src/components/layout/error-trigger/ErrorTrigger.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
-import commonStyles from '@/components/common.module.css';
-import styles from './ErrorTrigger.module.css';
+import commonStyles from '@/components/common.module.scss';
+import styles from './ErrorTrigger.module.scss';
 
 const ErrorTrigger: FC = () => {
   return (
