+++ src/components/layout/search/Search.module.scss
@@ -10,10 +10,8 @@
   position: fixed;
   width: 100%;
   background-color: #000000;
-}
 
-@media screen and (max-width: 900px) {
-  .search {
+  @media screen and (max-width: 900px) {
     padding-bottom: 45px;
   }
 }
@@ -23,13 +21,12 @@
   border: 0;
   height: 25px;
   padding-left: 5px;
-  font-weight: bold;
   background-color: #ffffff50;
   transition: background-color 0.15s ease-in-out;
   width: 50%;
   max-width: 500px;
-}
 
-.searchInput:focus {
-  background-color: #ffffff;
+  &:focus {
+    background-color: #ffffff;
+  }
 }
